agents/requirement-analysis-agent/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "requirement-analysis-agent"
version = "1.0.0"
description = "Requirement Analysis Agent - Analyzes requirements from multiple input formats, assesses quality, identifies gaps, and prepares structured output for test case generation"
authors = [
    {name = "QA Platform Team", email = "team@qaplatform.io"}
]
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "grpcio-reflection>=1.76.0",
    "grpcio-health-checking>=1.76.0",
    "anthropic>=0.75.0",
    "weaviate-client>=4.19.0",
    "python-dotenv>=1.2.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "structlog>=25.5.0",
    "aiohttp>=3.13.2",
    "prometheus-client>=0.23.1",
    "tenacity>=9.1.2",
    "jinja2>=3.1.6",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "black>=25.12.0",
    "ruff>=0.14.10",
    "mypy>=1.19.1",
    "types-protobuf>=6.32.1",
    "types-pyyaml>=6.0.12",
]

[project.scripts]
requirement-analysis-agent = "requirement_analysis_agent.main:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["requirement_analysis_agent*"]
namespaces = false

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/requirement_analysis_agent"]
omit = [
    "*/proto/*",
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
fail_under = 65  # External service clients (Weaviate, Anthropic, Domain Agent) require integration tests

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    proto
    | migrations
    | .venv
)/
'''

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "B905",  # `zip()` without an explicit `strict=` parameter
]
exclude = [
    ".venv",
    "proto",
    "__pycache__",
]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true
strict_equality = true
exclude = [
    "proto/",
    "tests/",
]
